[SH-89] 행동카드 사용 구현

+++ src/main/java/com/snowhite/server/repository/CardRepository.java
@@ -3,5 +3,5 @@ package com.snowhite.server.repository;
 import com.snowhite.server.domain.entity.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
 }
